CASMINST-3483: set CMN gateway "correctly"
"Correctly" is in quotes here because the fact that we're special-casing
CAN and CMN gateways here means the code is broken in a more fundamental
way.  The gateways are already set correctly before entering writeOutput()
and then seem to get re-templated incorrectly after that.  More
investigation is needed.  For now, copy what was done for the CAN.