CASMPET-4885 Run join-spire-on-storage script
This will attempt to join the storage nodes to spire the first time an
NCN is booted. It should fail until spire-server is running in the k8s
cluster. This is expected to succeed when ncn-m001 is rebooted. It will
not attempt to join nodes that already exist and contains a spire-server
healthcheck that times out after 5 minutes. This health check will fail
until sysmgmt is installed via loftsman.