Fix Signed RPM Builds Dying on ClamAV Permission Denied Errors (#95)
* Also always sign release/* branches

We have some old release branches which if rebuilt won't be signed
(however git-tags would work). This change would enable release branches
to always sign going forward, presuming we stick to existing
infrastructure and also need to revisit using such branches.

* Fix RPM signing ClamAV Failures

Always chown the buildroot since we run a portion of the build as root
and another as jenkins. This will ensure that files and directories are
usable on any subsequent build.