CASMINST-4117
Updated .version and handoff-bss-metadata.go to prevent leaks of m002 sun mac addresses into the m001 node. Also updated to ensure ifname renaming is done correctly if 2 NICs present. Removed bond interface configuration from buildPITArgs as we create the bond with cloud-init that will happen on boot of the node, after the csi hand off and don't want/need it here in the handoff-bss-metadata.go