MTL-1876 Private NCN Bucket Support
This drops support for uploading NCN images with CSI, CSI was using
special casing for uploading NCN artifacts into a public bucket which is
no longer needed. This also updates the CSI code to use a new
`metal.server` parameter which will be interpreted by BSS. BSS will
provide the NCN a server URL for pulling from a private bucket.