MTL-1971 Accept `1.4` `system_config.yaml` files
This value contols which `system_config.yaml` files are accepted,
forcing the user to update the file when going to a new CSM release.
Otherwise old values (deprecated, removed) and new values can be
accidentally skipped or set incorrectly between releases.

This effectively denies any config file that isn't for 1.4.