MTL-1476 Support GPT ISO and overlayFS (#279)
This includes support for the new LiveCD ISO being built from node-images
(instead of the  cray-pre-install-toolkit). The new ISO has a fixed GPT
structure, no longer showing up as MSDOS. This entails some
modifications to the `parted` commands.

This also has additional steps for installing the dependent directory
structure required for overlayFS to work properly. The new ISO uses
`dmsquash-live` instead of the kiwi dracut module for handling the
overlayFS, this entails needing certain directories to exist during the
boot. This is an important change for MTL-2035.