Cleanup (#292)
* Remove CSM 0.9 specific option

This option is out-of-date and unused. `system_config.yaml` has been
setting a blank string for `manifest-release` for a long time.

Removing this code shrinks our CSM 0.9 footprint, and removes old Stash
URLs for shasta-cfg.

* Run `go fmt`

* Pages workflow using wrong Go version

Was using Go 1.16 and failing, should be using Go 1.19 anyway to align
with the rest of the CSI codebase. Using 1.19 fixed the page workflow.

Use latest hugo version.

* Breakup workflows

* Build tweaks