Also publish to sp3 and sp4
There are a few too many doc steps that refer to a SLE based repository
for installing cray-site-init at this time.

To facilitate the transition period, continue publishing to SP3 and SP4
directories. This can still do without the matrix builds, and we can
just push the same RPM to all 3 places during this interim period.