MTL-2000 Needs top-level `Zypper` key
The `repos` key is actually nested under `zypper`, cloud-init
specifically looks for the parent key `zypper` for determining to run
the `zypper_add_repos` module.

This is compatible with the existing format of the `cloud-init.yaml`
file, which currently does not house anything under a `zypper` key.
Since this function won't be used after CSM 1.6, it's fine to leave this
way.