MTL-2000 Needs top-level `Zypper` key (#333)
* MTL-2000 Needs top-level `Zypper` key

The `repos` key is actually nested under `zypper`, cloud-init
specifically looks for the parent key `zypper` for determining to run
the `zypper_add_repos` module.

This is compatible with the existing format of the `cloud-init.yaml`
file, which currently does not house anything under a `zypper` key.
Since this function won't be used after CSM 1.6, it's fine to leave this
way.

* Fix help message

The help message for `csi patch packages` is confusing, it is hard to
differentiate between `--config` and `--config-file`. Instead of
changing the names, this just makes it more clear which one is the CSI
config file.

* `https://packages/` is hit during MTL

During a fresh install (before CSM), we will be calling the `zypper_add_repos` module
before `net-init.sh` runs when only `mtl` resolves in DNS.

During upgrades when CSM is installed, `packages` will resolve properly
to the 10.92 address.