CASMINST-7396 `Global.user-data` is `nil`
In some contexts, such as vShasta installs, the `user-data` is munged and not expected to be `nil`.

PR #488 changed the Basecamp datastructures from ad-hoc structs to `hms-bss.types.CloudInit`. As a side-effect, a new default value of `nil` was adopted for `Global.UserData`.

This change ensures that `Global.user-data` is not `nil`, making it an empty dictionary.