Update / Fix GitHub Page Workflow
The current GitHub pages workflow is broken and out-of-date.

This changes the GitHub pages workflow from GitHub Branch deployments to GitHub Actions deployments.

The pages workflow was updated to use Go 1.24 with Hugo 0.148.1, copying the new workflow suggestion from: https://gohugo.io/host-and-deploy/host-on-github-pages/.

The `hugo/content/commands` pages were removed and are now ignored in `.gitignore`, these files are generated on-build, and we do not need to maintain them in the repo.

`.gitignore` now ignores `docs/` and `.hugo_build.lock` files (generated/temp files from hugo).